pyproject.toml
[tool.poetry]
name = "muglegeonyang"
version = "0.1.0"
description = ""
authors = ["wkdsh21 <wkdsh21@naver.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.1.2"
tzdata = "^2024.2"
ipython = "^8.29.0"
django-extensions = "^3.2.3"
drf-yasg = "^1.21.8"
djangorestframework = "^3.15.2"
django-ninja = "^1.3.0"
mypy = "^1.13.0"
black = "^24.10.0"
isort = "^5.13.2"
django-stubs = "^5.1.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
